remove another iceberg config
