+++ src/moonlink_service/Cargo.toml
@@ -8,6 +8,7 @@ license = { workspace = true }
 arrow-ipc = { workspace = true }
 arrow-schema = { workspace = true }
 clap = { workspace = true }
+moonlink = { path = "../moonlink" }
 moonlink_backend = { path = "../moonlink_backend" }
 moonlink_metadata_store = { path = "../moonlink_metadata_store" }
 moonlink_rpc = { path = "../moonlink_rpc" }
