+++ src/moonlink_service/src/lib.rs
@@ -2,7 +2,7 @@ mod error;
 
 use arrow_ipc::writer::StreamWriter;
 pub use error::{Error, Result};
-use moonlink_backend::table_config::{TableConfig, TableCreationConfig};
+use moonlink_backend::table_creation_config::{TableConfig, TableCreationConfig};
 use moonlink_backend::MoonlinkBackend;
 use moonlink_metadata_store::SqliteMetadataStore;
 use moonlink_rpc::{read, write, Request, Table};
@@ -74,10 +74,9 @@ async fn handle_stream(
                     mooncake_creation_config: TableConfig {
                         enable_index_merge: true,
                     },
-                    storage_creation_config:
-                        moonlink_backend::table_config::IcebergStorageConfig::FsStorageConfig(
-                            backend.get_base_path(),
-                        ),
+                    storage_creation_config: moonlink::FileSystemConfig::FileSystem {
+                        root_directory: backend.get_base_path(),
+                    },
                 };
                 backend
                     .create_table(database_id, table_id, src, src_uri, table_creation_config)
